+++ src/wrappers/library.js
